+++ src/backend/src/routes/index.js
@@ -8,6 +8,7 @@ const bookingRoutes = require('./booking.routes');
 const notificationRoutes = require('./notification.routes')
 const mapRoutes = require('./maps.routes')
 const reviewRoutes = require('./review.routes')
+const reportRoutes = require('./report.routes')
 
 const router = express.Router();
 
@@ -19,6 +20,7 @@ router.use('/driver-verifications', driverVerifRoutes);
 router.use('/bookings', bookingRoutes);
 router.use('/notifications', notificationRoutes);
 router.use('/reviews', reviewRoutes);
+router.use('/reports', reportRoutes);
 router.use('/api/maps', mapRoutes);
 
 module.exports = router;